requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
pdfminer.six
pillow
pytesseract
requests
pyyaml
